% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_atlas.R
\name{as_brain_atlas}
\alias{as_brain_atlas}
\title{Create brain atlas}
\usage{
as_brain_atlas(x)
}
\arguments{
\item{x}{object to make into a brain_atlas}
}
\value{
an object of class 'brain_atlas'.
}
\description{
Coerce object into an object of class
'brain_atlas'.
}
