% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseg-atlas.R
\name{as_ggseg_atlas}
\alias{as_ggseg_atlas}
\alias{as_ggseg_atlas.default}
\alias{as_ggseg_atlas.data.frame}
\alias{as_ggseg_atlas.ggseg_atlas}
\alias{as_ggseg_atlas.brain_atlas}
\title{Create ggseg atlas}
\usage{
as_ggseg_atlas(x)

\method{as_ggseg_atlas}{default}(x)

\method{as_ggseg_atlas}{data.frame}(x)

\method{as_ggseg_atlas}{ggseg_atlas}(x)

\method{as_ggseg_atlas}{brain_atlas}(x)
}
\arguments{
\item{x}{object to make into a ggseg_atlas}
}
\value{
Object of class 'ggseg_atlas'
}
\description{
Create ggseg atlas
}
