% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlas.info}
\alias{atlas.info}
\title{Information on ggseg atlases}
\format{A data.frame summarising the ggseg data.}
\usage{
data(atlas.info)
}
\description{
Information on ggseg atlases
}
\keyword{data}
\keyword{summary}
