% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_palettes.R, R/data.R
\docType{data}
\name{atlas_info}
\alias{atlas_info}
\title{Get information about atlas}
\format{A tibble summarising the ggseg data.}
\usage{
atlas_info(name = NULL)

data(atlas_info)
}
\arguments{
\item{name}{name of atlas as character string. If left to NULL gives all.}
}
\value{
data.frame
}
\description{
Get information about atlas

Information on ggseg atlases
}
\keyword{data}
\keyword{summary}
