% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_atlas.R
\name{brain_atlas}
\alias{brain_atlas}
\title{Constructor for brain atlas}
\usage{
brain_atlas(atlas, type, data)
}
\arguments{
\item{atlas}{atlas short name, length one}

\item{type}{atlas type, cortical or subcortical, length one}

\item{data}{data.frame with atlas data}
}
\description{
Constructor for brain atlas
}
