% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_atlas.R
\name{brain_data-class}
\alias{brain_data-class}
\alias{brain_data}
\title{\code{brain_data} class}
\usage{
brain_data(x)
}
\arguments{
\item{x}{data.frame to be made a brain_data}
}
\value{
object of class brain_data, consisting of sf polygon
       data for brain atlas plotting.
}
\description{
\code{brain_data} class
}
