% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_palettes.R
\name{brain_pal}
\alias{brain_pal}
\title{Generate palettes from the ggseg atlases}
\usage{
brain_pal(name, n = "all", direction = 1, unname = FALSE)
}
\arguments{
\item{name}{String name of atlas}

\item{n}{Number of colours to return (or "all" [default])}

\item{direction}{Direction of  HEX, -1 reverses order}

\item{unname}{logical, if colours are to be unnamed before returning.
Neccessary if applying palette to other data than the brain atlas it comes from.}
}
\description{
\code{brain_pal} return HEX colours for the different ggseg atlases.
}
