% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_palettes.R, R/data.R
\docType{data}
\name{brain_pals_info}
\alias{brain_pals_info}
\title{Get info on brain palettes}
\format{A data.frame summarising the ggseg palettes.}
\usage{
brain_pals_info()

data(brain_pals_info)
}
\value{
data.frame
}
\description{
Get info on brain palettes

Information on ggseg atlas palettes
}
\keyword{palettes}
