% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_palettes.R
\name{display.brain.pal}
\alias{display.brain.pal}
\title{Plot the colours of the atlases for selection}
\usage{
display.brain.pal(name = "all", n = "all")
}
\arguments{
\item{name}{String name of atlas}

\item{n}{Number of colours to return (or "all" [default])}
}
\description{
\code{display.brain.pal} plots all the colours for each atlas.
}
