% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dk}
\alias{dk}
\title{Desikan-Killiany Cortical Atlas}
\format{
A data.frame with 10913 observations and 11 variables
}
\usage{
data(dk)
}
\description{
Coordinate data for the Desikan-Killany Cortical atlas,
with 40 regions in on the cortical surface of the brain.
}
\examples{
data(dk)
}
\references{
Fischl et al. (2004) Cerebral Cortex 14:11-22
(\href{https://academic.oup.com/cercor/article/14/1/11/433466}{PubMed})
}
\seealso{
Other ggseg_atlases: 
\code{\link{aseg}}
}
\concept{ggseg_atlases}
\keyword{datasets}
