% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dkt3d}
\alias{dkt3d}
\title{Desikan-Killiany Cortical Atlas}
\format{An object of class \code{list} of length 3.}
\usage{
data(dkt3d)
}
\description{
Coordinate data for the Desikan-Killany Cortical atlas,
with 40 regions in on the cortical surface of the brain.
}
\examples{
data(dkt3d)
}
\references{
Fischl et al. (2004) Cerebral Cortex 14:11-22
(\href{https://academic.oup.com/cercor/article/14/1/11/433466}{PubMed})
}
\keyword{datasets}
