% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_freesurfer_stats.R
\name{find_hemi_fromfile}
\alias{find_hemi_fromfile}
\title{helper function to easily grab hemisphere information from file path}
\usage{
find_hemi_fromfile(path)
}
\arguments{
\item{path}{file path}
}
\description{
helper function to easily grab hemisphere information from file path
}
