% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geom_brain}
\alias{geom_brain}
\title{Brain geom}
\usage{
geom_brain(
  mapping = aes(),
  data = NULL,
  atlas = NULL,
  stat = "sf",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{arguments to \code{\link[sfr]{geom_sf}}}
}
\value{

}
\description{
call to \code{\link[sfr]{geom_sf}}
}
