% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_palettes.R
\name{get_colours}
\alias{get_colours}
\title{Get colours from brain palettes}
\usage{
get_colours(name, n, unname, package)
}
\arguments{
\item{name}{String name of atlas}

\item{n}{Number of colours to return (or "all" [default])}

\item{unname}{logical, if colours are to be unnamed before returning.
Neccessary if applying palette to other data than the brain atlas it comes from.}

\item{package}{package to get brain_pals data from (ggseg or ggsegExtra)}
}
\description{
Needed to display the colours, and inspect
the number of colours of each palette
}
