% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_palettes.R
\name{get_pals}
\alias{get_pals}
\title{Get brain palettes}
\usage{
get_pals(package = "ggseg")
}
\arguments{
\item{package}{package to get brain_pals data from (ggseg or ggsegExtra)}
}
\description{
Brain palettes can be found in either ggseg
or ggsegExtra. This functions is a helper
to choose which package to get the palette
from.
}
