% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseg-package.R, R/ggseg.R
\docType{package}
\name{ggseg}
\alias{ggseg-package}
\alias{ggseg}
\title{ggseg: Visualise brain atlas data with ggplot2}
\usage{
ggseg(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
Does not return; always raises an error.
}
\description{
Provides a ggplot2 geom and position for visualising brain region data on
cortical, subcortical, and white matter tract atlases. Brain atlas
geometries are stored as simple features (sf), enabling seamless
integration with the ggplot2 ecosystem including faceting, custom scales,
and themes.

\code{ggseg()} is defunct as of version 2.0.0. Use
\code{ggplot() + geom_brain()} instead.
}
\details{
The main entry point is \code{\link[=geom_brain]{geom_brain()}}, which accepts a \code{ggseg_atlas}
object and optional user data. Use \code{\link[=position_brain]{position_brain()}} to control the
layout of brain slices/views.
}
\examples{
\dontrun{
ggseg()
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ggsegverse.github.io/ggseg/}
  \item \url{https://github.com/ggsegverse/ggseg}
  \item Report bugs at \url{https://github.com/ggsegverse/ggseg/issues}
}


\code{\link[=geom_brain]{geom_brain()}} for the replacement API.
}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})
  \item Didac Vidal-Piñeiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

Other contributors:
\itemize{
  \item Ramiro Magno \email{ramiro.magno@gmail.com} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID}) [contributor]
  \item Center for Lifespan Changes in Brain and Cognition, University of Oslo, Norway [copyright holder]
}

}
\keyword{internal}
