% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseg3d.R
\name{ggseg3d}
\alias{ggseg3d}
\title{Plot 3D brain parcellations}
\usage{
ggseg3d(.data = NULL, atlas = "dkt_3d", surface = "LCBC",
  hemisphere = c("right", "subcort"), label = "area", text = NULL,
  colour = "colour", palette = NULL, na.colour = "darkgrey",
  na.alpha = 1, remove.axes = TRUE, show.legend = TRUE,
  camera = "lateral")
}
\arguments{
\item{.data}{A data.frame to use for plot aesthetics. Must include a
column called "area" corresponding to areas.}

\item{atlas}{Either a string with the name of a 3d atlas to use.}

\item{surface}{String. Which surface to plot. Either "pial","white", or "inflated"[default]}

\item{hemisphere}{String. Hemisphere to plot. Either "left" or "right"[default],
can also be "subcort".}

\item{label}{String. Quoted name of column in atlas/data that should be used to name traces}

\item{text}{String. Quoated name of column in atlas/data that should be added as extra
information in the hover text.}

\item{colour}{String. Quoted name of column from which colour should be supplied}

\item{palette}{String. Either name of paletteer palette or vector of hex colours,
used if colour is numeric.}

\item{na.colour}{String. Either name, hex of RGB for colour of NA in colour.}

\item{na.alpha}{Numeric. A number between 0 and 1 to control transparency of NA-regions.}

\item{remove.axes}{Logical. Should axis and grid be removed.}

\item{show.legend}{Logical. Toggle legend if colour is numeric.}

\item{camera}{String of "medial" or "lateral", or list of x, y, and z positions
for initial camera position.}
}
\value{
a plotly object
}
\description{
\code{ggseg3d} plots and returns a plotly mesh3d object.
}
\details{
\describe{
\strong{Available atlases:}
  \item{`dkt_3d`}{
    The Desikan-Killiany Cortical Atlas [default], Freesurfer cortical segmentations, in 3dmesh format
  }
  \item{`yeo7_3d`}{
    Yeo 2011 7 resting-state networks, in 3dmesh format
  }
  \item{`yeo17_3d`}{
    Yeo 2011 17 resting-state networks, in 3dmesh format
  }
}

\strong{Available surfaces:}
\itemize{
\item `inflated` Fully inflated surface
\item `white` white matter surface
 }
}
\examples{
library(ggplot2)
ggseg3d()
ggseg3d(surface="white")
ggseg3d(surface="inflated")
ggseg3d(remove.axes = FALSE)

}
\seealso{
\code{\link[plotly]{plot_ly}}, \code{\link[plotly]{add_trace}}, \code{\link[plotly]{layout}}, the plotly package
}
\author{
Athanasia Mowinckel and Didac Pineiro
}
