% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseg_atlas.R
\name{ggseg_atlas-class}
\alias{ggseg_atlas-class}
\alias{as_ggseg_atlas}
\alias{ggseg_atlas}
\title{`ggseg_atlas` class}
\usage{
as_ggseg_atlas(
  x = data.frame(.long = double(), .lat = double(), .id = character(), area =
    as.character(), hemi = character(), side = character())
)
}
\arguments{
\item{x}{dataframe to be made a ggseg-atlas}
}
\description{
The `ggseg_atlas` class is a subclass of [`data.frame`][base::data.frame()],
created in order to have different default behaviour. It heavily relieas on
the "tibble" [`tbl_df`][tibble::tibble()].
[tidyverse](https://www.tidyverse.org/packages/), including
[dplyr](http://dplyr.tidyverse.org/),
[ggplot2](http://ggplot2.tidyverse.org/),
[tidyr](http://tidyr.tidyverse.org/), and
[readr](http://readr.tidyverse.org/).
}
\section{Properties of `ggseg_atlas`}{


Objects of class `ggseg_atlas` have:
* A `class` attribute of `c("ggseg_atlas", "tbl_df", "tbl", "data.frame")`.
* A base type of `"list"`, where each element of the list has the same
  [NROW()].
* Alot of this script and its functions are taken from the
  [`tibble`][tibble::tibble()]-package
}

\seealso{
[tibble()], [as_tibble()], [tribble()], [print.tbl()], [glimpse()]
}
