% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseg-atlas.R
\name{ggseg_atlas}
\alias{ggseg_atlas}
\title{`ggseg_atlas` class}
\usage{
ggseg_atlas(x)
}
\arguments{
\item{x}{data.frame to be made a ggseg-atlas}
}
\value{
a tibble with polygon coordinates for plotting brain regions
}
\description{
The `ggseg_atlas` class is a subclass of [`data.frame`][data.frame()],
created in order to have different default behavior. It heavily relies on
the "tibble" [`tbl_df`][tibble()].
[tidyverse](https://www.tidyverse.org/packages/), including
[dplyr](http://dplyr.tidyverse.org/),
[ggplot2](http://ggplot2.tidyverse.org/),
[tidyr](http://tidyr.tidyverse.org/), and
[readr](http://readr.tidyverse.org/).
}
\section{Properties of `ggseg_atlas`}{


Objects of class `ggseg_atlas` have:
* A `class` attribute of `c("ggseg_atlas", "tbl_df", "tbl", "data.frame")`.
* A base type of `"list"`, where each element of the list has the same
  [NROW()].
* A lot of this script and its functions are taken from the
  [`tibble`][tibble()]-package
}

