% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glasser}
\alias{glasser}
\title{Parcellation from the Human Connectome Project}
\format{A data.frame with 2483 observations and 8 variables
\describe{
  \item{long}{coordinates for the x-axis}
  \item{lat}{coordinates for the y-axis}
  \item{area}{name of region}
  \item{hemi}{dummy name of the hemisphere}
  \item{side}{which side to view (medial, lateral)}
  \item{label}{label name from Freesurfer}
}}
\usage{
glasser
}
\description{
Parcellation from the Human Connectome Project
}
\examples{
data(glasser)
}
\references{
Glasser et al. (2016) Nature, volume 536, pages 171–178
(\href{https://www.nature.com/articles/nature18933}{PubMed})
}
\keyword{datasets}
