% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_atlases}
\alias{install_atlases}
\title{Convenience function to install all atlases from ggsegExtra-repo}
\usage{
install_atlases(repo = "LCBC-UiO/ggsegExtra", ...)
}
\arguments{
\item{repo}{repository to install via github}

\item{...}{other options to install_github}
}
\description{
\code{install_atlases} calls devtools::install_github() with repo = "LCBC-UiO/ggsegExtra"
}
\author{
Athanasia Mowinckel and Didac Pineiro
}
