% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_atlas.R
\name{is_brain_atlas}
\alias{is_brain_atlas}
\title{Validate brain atlas}
\usage{
is_brain_atlas(x)
}
\arguments{
\item{x}{an object}
}
\value{
logical if object is of class 'brain_atlas'
}
\description{
Validate brain atlas
}
