% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jhu}
\alias{jhu}
\title{Parcellation from JHU}
\format{A data.frame with 11341 observations and 11 variables
\describe{
  \item{long}{coordinates for the x-axis}
  \item{lat}{coordinates for the y-axis}
  \item{area}{name of region}
  \item{hemi}{dummy name of the hemisphere}
  \item{side}{which side to view (sagittal)}
  \item{label}{label name from Freesurfer}
  \item{atlas}{name of the atlas}
}}
\usage{
jhu
}
\description{
Parcellation from JHU
}
\examples{
data(jhu)
}
\references{
Hua et al., Tract probability maps in stereotaxic spaces:
analysis of white matter anatomy and tract-specific quantification. NeuroImage, 39(1):336-347 (2008)
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2724595/}{PubMed})
}
\keyword{datasets}
