% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R, R/geom.R
\name{layer_brain}
\alias{layer_brain}
\title{Create a new sf layer that auto-maps geometry data}
\usage{
layer_brain(
  geom = NULL,
  stat = NULL,
  data = NULL,
  mapping = NULL,
  position = NULL,
  params = list(),
  inherit.aes = TRUE,
  check.aes = TRUE,
  check.param = TRUE,
  show.legend = NA
)

layer_brain(
  geom = NULL,
  stat = NULL,
  data = NULL,
  mapping = NULL,
  position = NULL,
  params = list(),
  inherit.aes = TRUE,
  check.aes = TRUE,
  check.param = TRUE,
  show.legend = NA
)

layer_brain(
  geom = NULL,
  stat = NULL,
  data = NULL,
  mapping = NULL,
  position = NULL,
  params = list(),
  inherit.aes = TRUE,
  check.aes = TRUE,
  check.param = TRUE,
  show.legend = NA
)
}
\description{
The `layer_sf()` function is a variant of [`layer()`] meant to be used by
extension developers who are writing new sf-based geoms or stats.
The sf layer checks whether the data contains a geometry column, and
if one is found it is automatically mapped to the `geometry` aesthetic.

The `layer_brain()` function is a variant of [`layer_sf()`] meant to be used by
extension developers who are writing new sf-based geoms or stats.
The sf layer checks whether the data contains a geometry column, and
if one is found it is automatically mapped to the `geometry` aesthetic.

The `layer_sf()` function is a variant of [`layer()`] meant to be used by
extension developers who are writing new sf-based geoms or stats.
The sf layer checks whether the data contains a geometry column, and
if one is found it is automatically mapped to the `geometry` aesthetic.
}
\keyword{internal}
