% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{midsagittal}
\alias{midsagittal}
\title{Parcellation from the Human Connectome Project}
\format{A data.frame with 2483 observations and 8 variables
\describe{
  \item{long}{coordinates for the x-axis}
  \item{lat}{coordinates for the y-axis}
  \item{area}{name of region}
  \item{hemi}{dummy name of the hemisphere}
  \item{side}{which side to view (sagittal)}
  \item{label}{label name from Freesurfer}
  \item{atlas}{name of the atlas}
}}
\usage{
midsagittal
}
\description{
Parcellation from the Human Connectome Project
}
\examples{
data(midsagittal)
}
\keyword{datasets}
