% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_atlas.R
\name{plot.brain_atlas}
\alias{plot.brain_atlas}
\title{Plot a brain atlas}
\usage{
\method{plot}{brain_atlas}(x, ...)
}
\arguments{
\item{x}{A `brain_atlas` object.}

\item{...}{Additional arguments passed to [geom_brain()].}
}
\value{
A [ggplot2::ggplot] object.
}
\description{
Plots a brain atlas using [geom_brain()], coloured by region label. If the
atlas contains a palette, it is applied via [ggplot2::scale_fill_manual()].
}
\examples{
plot(dk)
plot(aseg)
}
