% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeo17}
\alias{yeo17}
\title{Yeo 17 Resting-state Cortical Parcellations}
\format{A data.frame with 8203 observations and 10 variables
\describe{
  \item{long}{coordinates for the x-axis}
  \item{lat}{coordinates for the y-axis}
  \item{area}{name of network}
  \item{hemi}{name of the hemisphere (left, right)}
  \item{side}{which side to view (medial, lateral)}
  \item{network}{network number (1:17)}
  \item{label}{unique name to each node}
}}
\usage{
data(yeo17)
}
\description{
Coordinate data for the resting-state networks of
the Yeo 2011 17 networks.
}
\examples{
data(yeo17)
}
\references{
Yeo et al. (2011) J. Neurophysiology 16(3):1125-1165
(\href{https://www.ncbi.nlm.nih.gov/pubmed/21653723}{PubMed})
}
\keyword{datasets}
