% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeo7_3d}
\alias{yeo7_3d}
\title{Yeo 7 Resting-state Cortical Parcellations}
\format{A tibble with 4 observations and a nested data.frame
\describe{
  \item{surf}{type of surface (`inflated` or `white`)}
  \item{hemi}{hemisphere (`left`` or `right`)}
  \item{data}{data.frame of necessary variables for plotting
  }

  \item{atlas}{String. atlas name}
  \item{roi}{numbered region from surface}
  \item{annot}{concatenated region name}
  \item{label}{label `hemi_annot` of the region}
  \item{mesh}{list of meshes in two lists: vb and it}
  \item{area}{name of area in full}
  \item{colour}{HEX colour of region}
}}
\usage{
data(yeo7_3d)
}
\description{
Mesh data for the resting-state networks of
the Yeo 2011 7 networks.
}
\examples{
data(yeo7_3d)
}
\references{
Yeo et al. (2011) J. Neurophysiology 16(3):1125-1165
(\href{https://www.ncbi.nlm.nih.gov/pubmed/21653723}{PubMed})
}
\keyword{datasets}
