% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aseg}
\alias{aseg}
\title{Freesurfer automatic subcortical segmentation of a brain volume}
\format{
An object of class \code{brain_atlas} of length 4.
}
\usage{
data(aseg)
}
\value{
An object of class 'brain_atlas' for plotting with ggseg
}
\description{
Coordinate data for the subcortical parcellations implemented
in Freesurfer.
}
\examples{
data(aseg)
}
\references{
Fischl et al., (2002). Neuron, 33:341-355
(\href{https://www.pubmed.ncbi.nlm.nih.gov/11832223}{PubMed})
}
\seealso{
Other ggseg_atlases: 
\code{\link{dk}}
}
\concept{ggseg_atlases}
\keyword{datasets}
