% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_info.R
\name{atlas_info}
\alias{atlas_info}
\title{Get information about atlas}
\usage{
atlas_info(name = NULL)
}
\arguments{
\item{name}{name of atlas as character string. If left to NULL gives all.}
}
\value{
data.frame
}
\description{
Get information about atlas
}
