% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlas_info}
\alias{atlas_info}
\title{Information on ggseg atlases}
\format{A tibble summarising the ggseg data.}
\usage{
data(atlas_info)
}
\description{
Information on ggseg atlases
}
\keyword{data}
\keyword{summary}
