% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brain.pal.info}
\alias{brain.pal.info}
\title{Information on ggseg atlas palettes}
\format{A data.frame summarising the ggseg palettes.}
\usage{
data(brain.pal.info)
}
\description{
Information on ggseg atlas palettes
}
\keyword{palettes}
