% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_atlas.R
\name{brain_atlas}
\alias{brain_atlas}
\title{Constructor for brain atlas}
\usage{
brain_atlas(atlas, type, data, palette = NULL)
}
\arguments{
\item{atlas}{atlas short name, length one}

\item{type}{atlas type, cortical or subcortical, length one}

\item{data}{data.frame with atlas data}

\item{palette}{named character vector of colours}
}
\value{
an object of class 'brain_atlas' containing information
       on atlas name, type, data and palette. To be used in plotting
       with \code{\link{geom_brain}}.
}
\description{
Creates an object of class 'brain_atlas'
that is compatible for plotting using the
ggseg-package plot functions
}
