% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_atlas.R
\name{brain_data-class}
\alias{brain_data-class}
\alias{brain_data}
\title{`brain_data` class}
\usage{
brain_data(
  x = data.frame(atlas = character(), type = character(), region = character(), hemi =
    character(), side = character(), ggseg = character(), geometry = character())
)
}
\arguments{
\item{x}{dataframe to be made a brain_data}
}
\description{
`brain_data` class
}
