% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_atlas.R
\name{brain_data-class}
\alias{brain_data-class}
\alias{brain_data}
\title{\code{brain_data} class}
\usage{
brain_data(
  x = data.frame(atlas = character(), type = character(), region = character(), hemi =
    character(), side = character(), geometry = character())
)
}
\arguments{
\item{x}{data.frame to be made a brain_data}
}
\description{
\code{brain_data} class
}
