% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_palettes.R
\name{brain_pal}
\alias{brain_pal}
\title{Generate palettes from the ggseg atlases}
\usage{
brain_pal(name, n = "all", direction = 1, unname = FALSE, package = "ggseg")
}
\arguments{
\item{name}{String name of atlas}

\item{n}{Number of colours to return (or "all" [default])}

\item{direction}{Direction of  HEX, -1 reverses order
Necessary if applying palette to other data than the brain atlas it comes from.}

\item{unname}{return unnamed vector (default = FALSE)}

\item{package}{package to get brain_pals data from (ggseg or ggsegExtra)}
}
\value{
vector of colours
}
\description{
\code{brain_pal} return HEX colours for the different ggseg atlases.
}
\examples{
brain_pal("dk")
brain_pal("aseg")
}
