% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_palettes.R
\name{brain_pals_info}
\alias{brain_pals_info}
\title{Get info on brain palettes}
\usage{
brain_pals_info(package = "ggseg")
}
\arguments{
\item{package}{package to get brain_pals data from (ggseg or ggsegExtra)}
}
\value{
data.frame with palette information
}
\description{
Get info on brain palettes
}
\examples{
brain_pals_info()
}
