% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_palettes.R
\name{display_brain_pal}
\alias{display_brain_pal}
\title{Plot the colours of the atlases for selection}
\usage{
display_brain_pal(name = "all", n = "all")
}
\arguments{
\item{name}{String name of atlas}

\item{n}{Number of colours to return (or "all" [default])}
}
\description{
\code{display_brain_pal} plots all the colours for each atlas.
}
