% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dk}
\alias{dk}
\title{Desikan-Killiany Cortical Atlas}
\format{
An object of class \code{brain_atlas} of length 4.
}
\usage{
data(dk)
}
\value{
An object of class 'brain_atlas' for plotting with ggseg
}
\description{
Coordinate data for the Desikan-Killiany Cortical atlas,
with 40 regions in on the cortical surface of the brain.
}
\examples{
data(dk)
}
\references{
Fischl et al. (2004) Cerebral Cortex 14:11-22
(\doi{10.1093/cercor/bhg087})
}
\seealso{
Other ggseg_atlases: 
\code{\link{aseg}}
}
\concept{ggseg_atlases}
\keyword{datasets}
