% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_freesurfer_stats.R
\name{find_subject_fromdir}
\alias{find_subject_fromdir}
\title{helper function to easily grab subject information from directory path}
\usage{
find_subject_fromdir(path)
}
\arguments{
\item{path}{file path}
}
\description{
helper function to easily grab subject information from directory path
}
