% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-brain.R
\docType{data}
\name{geom_brain}
\alias{geom_brain}
\alias{GeomBrain}
\alias{geom_brain_label}
\alias{geom_brain_text}
\title{Brain geom}
\usage{
geom_brain(
  mapping = aes(),
  data = NULL,
  atlas = NULL,
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

geom_brain_label(
  mapping = aes(),
  data = NULL,
  atlas = dk,
  stat = "sf_coordinates",
  position = "identity",
  ...,
  parse = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  label.padding = unit(0.25, "lines"),
  label.r = unit(0.15, "lines"),
  label.size = 0.25,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  fun.geometry = NULL
)

geom_brain_text(
  mapping = aes(),
  data = NULL,
  atlas = dk,
  stat = "sf_coordinates",
  position = "identity",
  ...,
  parse = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  check_overlap = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  fun.geometry = NULL
)
}
\arguments{
\item{mapping}{argument to pass to \code{\link[ggplot2]{aes}} to map
variables from the supplied data to the plot}

\item{data}{data.frame with data to plot}

\item{atlas}{object of type brain_atlas to plot}

\item{position}{position of the data. Default is "identity" but can be
changed by \code{\link{position_brain}}.}

\item{show.legend}{logical. Should legend be added or not.}

\item{inherit.aes}{logical. if aes should be inherited from the
main ggplot call or not}

\item{...}{arguments to \code{\link[ggplot2]{geom_sf}}}
}
\value{
ggplot object#'
}
\description{
call to \code{\link[ggplot2]{geom_sf}}
}
\keyword{datasets}
