% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseg-package.R, R/ggseg.R
\docType{package}
\name{ggseg}
\alias{ggseg}
\alias{ggseg-package}
\title{ggseg: Plot brain segmentations with ggplot and plotly}
\usage{
ggseg(data = NULL, atlas = "dkt", plot.areas = NULL,
  position = "dispersed", view = NULL, hemisphere = NULL,
  adapt_scales = TRUE, ...)
}
\arguments{
\item{data}{A data.frame to use for plot aesthetics. Should include a
column called "area" corresponding to aparc areas.}

\item{atlas}{Either a string with the name of atlas to use,
or a data.frame containing atlas information (i.e. pre-loaded atlas).}

\item{plot.areas}{Character vector, plots only areas specified in the vector.}

\item{position}{String choosing how to view the data. Either "dispersed"[default] or "stacked".}

\item{view}{String to choose view of the data. Any of c("lateral","medial")[default].}

\item{hemisphere}{String to choose hemisphere to plot. Any of c("left","right")[default].}

\item{adapt_scales}{if \code{TRUE}, then the axes will
be hemisphere without ticks.  If \code{FALSE}, then will be latitude
longitude values.  Also affected by \code{position} argument}

\item{...}{other options sent to ggplot2::geom_polygon for plotting, including
mapping aes (cannot include x, y, and group aethetics).}
}
\value{
a ggplot object
}
\description{
Plotting results from analyses done on data derived from brain
segmentations is a common need, but may be quite laborious.
Results from such analyses are usually easier to interpret if
the plot can mimmick the shape and position in the brain it
represents.
This package contains data from various brain parcellations,
with convenient functions to inspect the results directly on
a brain-plot wither with ggplot (2D polygons) or plotly (3d meshes).

\code{ggseg} plots and returns a ggplot object of plotted
aparc areas.
}
\details{
\describe{

\item{`dkt`}{
The Desikan-Killiany Cortical Atlas [default], Freesurfer cortical segmentations.}

\item{`yeo7`}{
Seven resting-state networks from Yeo et al. 2011, J. Neurophysiology}

\item{`yeo17`}{
Seventeen resting-state networks from Yeo et al. 2011, J. Neurophysiology}

\item{`aseg`}{
Freesurfer automatic subcortical segmentation of a brain volume}

}
}
\examples{
library(ggplot2)
ggseg()
ggseg(mapping=aes(fill=area))
ggseg(colour="black", size=.7, mapping=aes(fill=area)) + theme_void()
ggseg(atlas="yeo7")
ggseg(adapt_scales = FALSE, position = "stacked")
ggseg(adapt_scales = TRUE, position = "stacked")
ggseg(adapt_scales = TRUE)
ggseg(adapt_scales = FALSE)

}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{geom_polygon}},\code{\link[ggplot2]{coord_fixed}}
}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no}

Authors:
\itemize{
  \item Didac Vidal Pineiro \email{d.v.pineiro@psykologi.uio.no}
}


Athanasia Mowinckel and Didac Pineiro
}
\keyword{internal}
