% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseg.R
\name{ggseg}
\alias{ggseg}
\title{Plot brain parcellations}
\usage{
ggseg(data = NULL, atlas = "dkt", plot.areas = NULL,
  position = "dispersed", view = c("lateral", "medial", "axial",
  "sagittal"), hemisphere = c("right", "left"), adapt.scales = TRUE,
  ...)
}
\arguments{
\item{data}{A data.frame to use for plot aesthetics. Must include a
column called "area" corresponding to aparc areas.}

\item{atlas}{Either a string with the name of atlas to use,
or a data.frame containing atlas information (i.e. pre-loaded atlas).}

\item{plot.areas}{Character vector, plots only areas specified in the vector.}

\item{position}{String choosing how to view the data. Either "dispersed"[default] or "stacked".}

\item{view}{String to choose view of the data. Any of c("lateral","medial")[default].}

\item{hemisphere}{String to choose hemisphere to plot. Any of c("left","right")[default].}

\item{adapt.scales}{if \code{TRUE}, then the axes will
be hemisphere without ticks.  If \code{FALSE}, then will be latitude
longitude values.  Also affected by \code{position} argument}

\item{...}{other options sent to ggplot2::geom_poly for plotting}

\item{mapping}{ggplot2::geom_poly aethetics (cannot include x and y aethetics)}
}
\value{
a ggplot object
}
\description{
\code{ggseg} plots and returns a ggplot object of plotted
aparc areas.
}
\details{
\describe{

\item{`dkt`}{
The Desikan-Killiany Cortical Atlas [default], Freesurfer cortical segmentations.}

\item{`yeo7`}{
Seven resting-state networks from Yeo et al. 2011, J. Neurophysiology}

\item{`aseg`}{
Freesurfer automatic subcortical segmentation of a brain volume}
}
}
\examples{
library(ggplot2)
ggseg()
ggseg(na.fill = "transparent",mapping=aes(fill=area))
ggseg(colour="black", size=.7, mapping=aes(fill=area)) + theme_void()
ggseg(atlas="yeo7")
ggseg(adapt.scales = FALSE ,position = "stacked")
ggseg(adapt.scales = TRUE ,position = "stacked")
ggseg(adapt.scales = TRUE)
ggseg(adapt.scales = FALSE)

}
\seealso{
[ggplot()], [aes()], [geom_polygon()], [coord_fixed()] from the ggplot2 package
}
\author{
Athanasia Mowinckel and Didac Pineiro
}
