% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseg3d.R
\name{ggseg3d}
\alias{ggseg3d}
\title{Plot 3D brain parcellations}
\usage{
ggseg3d(data = NULL, atlas = "dkt3d", surface = "inflated",
  hemisphere = "right", label = "area", text = NULL,
  colour = "colour", palette = NULL, na.color = "darkgrey",
  remove.axes = TRUE, show.legend = TRUE, camera = list(x = 1.5, y =
  0, z = 1))
}
\arguments{
\item{data}{A data.frame to use for plot aesthetics. Must include a
column called "area" corresponding to areas.}

\item{atlas}{Either a string with the name of a 3d atlas to use.}

\item{surface}{String. Which surface to plot. Either "pial","white", or "inflated"[default]}

\item{hemisphere}{String. Hemisphere to plot. Either "left" or "right"[default].}

\item{label}{String. Quoted name of column in atlas/data that should be used to name traces}

\item{text}{String. Quoated name of column in atlas/data that should be added as extra
information in the hover text.}

\item{colour}{String. Quoted name of column from which colour should be supplied}

\item{palette}{String. Either name of paletteer palette or vector of hex colours,
used if colour is numeric.}

\item{na.color}{String. Either name, hex of RGB for colour of NA in colour.}

\item{remove.axes}{Logical. Should axis and grid be removed.}

\item{show.legend}{Logical. Toggle legend if colour is numeric.}

\item{camera}{list of x, y, and z positions for initial camera position.}
}
\value{
a plotly object
}
\description{
\code{ggseg3d} plots and returns a plotly mesh3d object.
}
\details{
\describe{
\item{`dkt3d`}{
The Desikan-Killiany Cortical Atlas [default], Freesurfer cortical segmentations, in 3dmesh format}
}
}
\examples{
library(ggplot2)
ggseg3d()
ggseg3d(surface="pial")
ggseg3d(remove.axes = FALSE)

}
\seealso{
[ggplot()], [aes()], [geom_polygon()], [coord_fixed()] from the ggplot2 package
}
\author{
Athanasia Mowinckel and Didac Pineiro
}
