% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseg-atlas.R
\name{is_ggseg_atlas}
\alias{is_ggseg_atlas}
\title{Validate ggseg_atlas}
\usage{
is_ggseg_atlas(x)
}
\arguments{
\item{x}{an object}
}
\value{
logical if object is of class 'ggseg_atlas'
}
\description{
Validate ggseg_atlas
}
