% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseg_atlas.R
\name{is_ggseg_atlas}
\alias{is_ggseg_atlas}
\title{Check if is ggseg_atlas-class}
\usage{
is_ggseg_atlas(x)
}
\arguments{
\item{x}{atlas object to check}
}
\value{
logical
}
\description{
Check if is ggseg_atlas-class
}
