% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{midsagittal}
\alias{midsagittal}
\title{Mid-sagittal Parcellations}
\format{A data.frame with 2483 observations and 8 variables}
\usage{
data(midsagittal)
}
\description{
Coordinate data for important structures in the mid-line
sagittal view of the brain
}
\examples{
data(midsagittal)
}
\keyword{(medial,}
\keyword{\describe{}
\keyword{\item{area}{name}
\keyword{\item{hemi}{dummy}
\keyword{\item{lat}{coordinates}
\keyword{\item{long}{coordinates}
\keyword{\item{side}{which}
\keyword{datasets}
\keyword{for}
\keyword{hemisphere}}
\keyword{lateral)}}
\keyword{name}
\keyword{of}
\keyword{region}}
\keyword{side}
\keyword{the}
\keyword{to}
\keyword{view}
\keyword{x-axis}}
\keyword{y-axis}}
\keyword{}}
