% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{midsagittal}
\alias{midsagittal}
\title{Mid-sagittal Parcellations}
\format{A data.frame with 2483 observations and 8 variables
\describe{
  \item{long}{coordinates for the x-axis}
  \item{lat}{coordinates for the y-axis}
  \item{area}{name of region}
  \item{hemi}{dummy name of the hemisphere}
  \item{side}{which side to view (medial, lateral)}
}}
\usage{
midsagittal
}
\description{
Coordinate data for important structures in the mid-line
sagittal view of the brain
}
\examples{
data(midsagittal)
}
\keyword{datasets}
