% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-brain.R
\name{position_brain}
\alias{position_brain}
\title{Alter brain atlas position}
\usage{
position_brain(position = "horizontal")
}
\arguments{
\item{position}{formula describing the rows ~ columns organization.}
}
\value{
a ggproto object
}
\description{
Function to be used in the position argument in geom_brain
to alter the position of the brain slice/views.
}
\examples{
library(ggplot2)
ggplot() +
  geom_brain(atlas = dk, aes(fill = region),
             position = position_brain(. ~ side + hemi ),
             show.legend = FALSE)

ggplot() +
  geom_brain(atlas = dk, aes(fill = region),
             position = position_brain(side ~ hemi ),
             show.legend = FALSE)
}
