% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseg-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dk}
\alias{aseg}
\alias{tracula}
\alias{suit}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggseg.formats}{\code{\link[ggseg.formats]{aseg}}, \code{\link[ggseg.formats]{dk}}, \code{\link[ggseg.formats]{suit}}, \code{\link[ggseg.formats]{tracula}}}
}}

