% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-brain.R
\name{scale_colour_brain}
\alias{scale_colour_brain}
\alias{scale_fill_brain}
\title{Colour scales from the ggbrain atlases}
\usage{
scale_colour_brain(palette = "dkt", n = "all", unname = F,
  direction = 1, aesthetics = "colour")

scale_fill_brain(palette = "dkt", n = "all", unname = F,
  direction = 1, aesthetics = "fill")
}
\description{
The `brain` scales provides colours for the different atlases in the
package, according to the colours used in the papers where the
atlases where first introduced.
}
\section{Palettes}{

The following palettes are available for use with these scales:
\describe{
  \item{Qualitative}{dkt, yeo7, yeo17, aseg}
}
}

\concept{colour scales}
