% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-brain.R
\name{scale_colour_brain}
\alias{scale_colour_brain}
\alias{scale_fill_brain}
\alias{scale_brain}
\title{Colour scales from the ggseg atlases}
\usage{
scale_colour_brain(name = "dkt", ...)

scale_colour_brain(name = "dkt", ...)

scale_fill_brain(name = "dkt", ...)

scale_brain(name = "dkt", ..., aesthetics = c("colour", "color",
  "fill"))
}
\arguments{
\item{name}{String name of atlas}

\item{...}{additional arguments to pass to \code{\link{brain_pal}}}

\item{aesthetics}{You can scale the brain mor generally with \code{scale_brain}
and a switch of the aesthetics.}
}
\description{
The `brain` scales provides colours for the different atlases in the
package, according to the colours used in the papers where the
atlases where first introduced.
}
\section{Palettes}{

The following palettes are available for use with these scales:
\describe{
  \item{Qualitative}{dkt, yeo7, yeo17, aseg}
}
}

\examples{
scale_brain()
scale_colour_brain()
scale_fill_brain()
}
