% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-brain.R
\name{scale_brain2}
\alias{scale_brain2}
\alias{scale_colour_brain2}
\alias{scale_color_brain2}
\alias{scale_fill_brain2}
\alias{scale_brain2-deprecated}
\title{Deprecated scale functions}
\usage{
scale_brain2(...)

scale_colour_brain2(...)

scale_color_brain2(...)

scale_fill_brain2(...)
}
\arguments{
\item{...}{Arguments passed to the replacement function.}
}
\value{
A ggplot2 scale object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions have been renamed for clarity:
\itemize{
\item \code{scale_brain2()} -> \code{\link[=scale_brain_manual]{scale_brain_manual()}}
\item \code{scale_fill_brain2()} -> \code{\link[=scale_fill_brain_manual]{scale_fill_brain_manual()}}
\item \code{scale_colour_brain2()} -> \code{\link[=scale_colour_brain_manual]{scale_colour_brain_manual()}}
\item \code{scale_color_brain2()} -> \code{\link[=scale_color_brain_manual]{scale_color_brain_manual()}}
}
}
\examples{
pal <- c("transverse temporal" = "#FF0000", "insula" = "#00FF00")
suppressWarnings(scale_fill_brain_manual(palette = pal))
}
