% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-brain.R
\name{scale_brain2}
\alias{scale_brain2}
\alias{scale_colour_brain2}
\alias{scale_color_brain2}
\alias{scale_fill_brain2}
\alias{scale_brain2-deprecated}
\title{Deprecated scale functions}
\usage{
scale_brain2(...)

scale_colour_brain2(...)

scale_color_brain2(...)

scale_fill_brain2(...)
}
\arguments{
\item{...}{Arguments passed to the replacement function.}
}
\description{
`r lifecycle::badge("deprecated")`

These functions have been renamed for clarity:
- `scale_brain2()` -> [scale_brain_manual()]
- `scale_fill_brain2()` -> [scale_fill_brain_manual()]
- `scale_colour_brain2()` -> [scale_colour_brain_manual()]
- `scale_color_brain2()` -> [scale_color_brain_manual()]
}
