% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-brain.R
\name{theme_brain}
\alias{theme_brain}
\alias{theme_darkbrain}
\alias{theme_custombrain}
<<<<<<< HEAD:man/theme_brain.Rd
\alias{theme_brain2}
=======
>>>>>>> 9330d878f2fd8bca2b91eab6cb80021f5e3d370a:man/ggtheme.Rd
\title{ggseg plot theme}
\usage{
theme_brain(text.size = 12, text.family = "mono")

theme_darkbrain(text.size = 12, text.family = "mono")

theme_custombrain(
  plot.background = "white",
  text.colour = "darkgrey",
  text.size = 12,
  text.family = "mono"
)
<<<<<<< HEAD:man/theme_brain.Rd

theme_brain2(
  plot.background = "white",
  text.colour = "darkgrey",
  text.size = 12,
  text.family = "mono"
)
=======
>>>>>>> 9330d878f2fd8bca2b91eab6cb80021f5e3d370a:man/ggtheme.Rd
}
\arguments{
\item{text.size}{Specify size of plot text}

\item{text.family}{Speficy font family}

\item{plot.background}{Specify fill of plot background (`theme_custombrain` only)}

\item{text.colour}{Speficy colour of plot text}
}
\description{
<<<<<<< HEAD:man/theme_brain.Rd
a set of themes created for the ggseg plots. Use theme() to tweak.
=======
a set of themes created for the ggseg plots. Use ggplot2::theme() to tweak.
>>>>>>> 9330d878f2fd8bca2b91eab6cb80021f5e3d370a:man/ggtheme.Rd
}
\details{
\describe{

\item{`theme_brain`}{
Default theme for ggseg. Transparent background, no axis lines, and no grid.}

\item{`theme_darkbrain`}{
Dark equivalent to theme_brain, with black background, and light text.}

\item{`theme_custombrain`}{
Theme for easy customisation of the brain themes.}

}
}
\seealso{
[ggplot()], [aes()], [geom_polygon()], [coord_fixed()] from the ggplot2 package
}
\author{
Athanasia Mowinckel
}
